+++ src/Ui/Bundle/EventBundle/Controller/CatalogController.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Ui\Bundle\EventBundle\Controller;
 
 use Proximum\Vimeet\Application\Command\Sheet\SheetViewed\Add;
@@ -51,8 +43,6 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
- * Class CatalogController
- *
  * Routes are being protected by security access checker
  *
  * @see CatalogAccessEventListener
@@ -109,7 +99,7 @@ class CatalogController extends Controller
             return $this->redirect($availabilityConfirmation->redirectRoute);
         }
 
-        if (!$sheet->isInCatalog()) {
+        if (!$sheet->isInInternalCatalog()) {
             throw $this->createAccessDeniedException('Sheet not in catalog');
         }